feat(deps): [TMP] update `kurigram` to `4db3f39`
I've had a lot of request timeouts recently, and it got much worse after
updating to 2.2.x. And I'm not the only one with this problem.
From what I understand, the problem got resolved in the `dev` branch.
So, I'm hardcoding the library to the latest current commit.
I really hope this fixes my problem.

Signed-off-by: Farzin Kazemzadeh <itisFarzin@gmail.com>